moved ui settings to one file
