+++ aiohttp_swagger3/__init__.py
@@ -1,14 +1,13 @@
 __all__ = (
+    "ReDocUiSettings",
     "SwaggerDocs",
     "SwaggerFile",
     "SwaggerUiSettings",
-    "ReDocUiSettings",
     "__version__",
 )
 __version__ = "0.3.5"
 __author__ = "Valetov Konstantin"
 
-from .redoc_ui_settings import ReDocUiSettings
 from .swagger_docs import SwaggerDocs
 from .swagger_file import SwaggerFile
-from .swagger_ui_settings import SwaggerUiSettings
+from .ui_settings import ReDocUiSettings, SwaggerUiSettings
